Little bit more of refactoring
