+++ near-bindgen-core/src/info_extractor/impl_item_method_info.rs
@@ -5,26 +5,20 @@ use syn::{ImplItemMethod, Type, Visibility};
 pub struct ImplItemMethodInfo {
     /// Information on the attributes and the signature of the method.
     pub attr_signature_info: AttrSigInfo,
-    /// Whether method has `pub` modifier or a part of trait implementation.
+    /// Whether method has `pub` modifier.
     pub is_public: bool,
-    /// The original code of the method.
-    pub original: ImplItemMethod,
     /// The type of the contract struct.
     pub struct_type: Type,
 }
 
 impl ImplItemMethodInfo {
     /// Process the method and extract information important for near-bindgen.
-    pub fn new(
-        original: ImplItemMethod,
-        struct_type: Type,
-        is_trait_impl: bool,
-    ) -> syn::Result<Self> {
+    pub fn new(original: ImplItemMethod, struct_type: Type) -> syn::Result<Self> {
         let attr_signature_info = AttrSigInfo::new(original.attrs.clone(), original.sig.clone())?;
         let is_public = match original.vis {
             Visibility::Public(_) => true,
-            _ => is_trait_impl,
+            _ => false,
         };
-        Ok(Self { attr_signature_info, is_public, original, struct_type })
+        Ok(Self { attr_signature_info, is_public, struct_type })
     }
 }
